ahora se comprobara que exista el directorio antes de eliminarlo

+++ model/publis/fotos/imgCache.php
@@ -19,11 +19,14 @@
         }
     }
     function rmTree($dir) {
-        $files = array_diff(scandir($dir), array('.','..'));
-        foreach ($files as $file)
-            if (is_dir("$dir/$file")) 
-                 rmTree("$dir/$file");
-            else unlink("$dir/$file");
-        return rmdir($dir);
+        if (file_exists($dir)){
+            $files = array_diff(scandir($dir), array('.','..'));
+            foreach ($files as $file)
+                if (is_dir("$dir/$file")) 
+                     rmTree("$dir/$file");
+                else unlink("$dir/$file");
+            return rmdir($dir);
+        }
+        return true;
     }
 ?>
